pyproject.toml
[tool.poetry]
name = "movie-pipeline-segments-validator"
version = "0.1.0"
description = "A simple GUI to validate detected segments and to generate edit decision files for movie-pipeline"
authors = ["GdPaul1234 <paul.godin1234@outlook.fr>"]
readme = "README.md"
packages = [{include = "movie_pipeline_segments_validator"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"
ffmpeg-python = "^0.2.0"
schema = "^0.7.5"
rich = "^13.6.0"
deffcode = "^0.2.5"
PySimpleGUI = "^4.60.5"
decoratoroperations = {git = "https://github.com/salesforce/decorator-operations"}
Pillow = "^10.1.0"
pydantic = {extras = ["dotenv"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
